+++ libs/feature/src/lib/services/http/files/content-file/index.ts
@@ -1,0 +1,1 @@
+export * from './create-content-file';
