--- libs/feature/src/lib/services/http/files/content-video/index.ts
@@ -1,1 +1,0 @@
-export * from './create-content-video';
