+++ libs/feature/src/lib/services/http/files/index.ts
@@ -1,1 +1,1 @@
-export * from './content-video';
+export * from './content-file';
